refactor: tags component の命名を PascalCase に統一
color-palette-picker を export 実体に合わせて ColorPaletteMenuItems に、
tag-create/merge/rename-modal を Tag*Modal に改称し、GlobalTag*Modal の
import・barrel export・stories/test を追随させる。

Refs #1530
