+++ apps/product/src/features/tags/components/ColorPaletteMenuItems.tsx
